Add the configuration option [plugin][load_path]
This option can be used to specify absolute and relative paths on which
the bot will looking for additional plugins.

The introduction of this feature will decreare the need for user to fork
this repo just to add some vertical/not-shareble plugins.
You can see some example scouting the repo network:
 https://github.com/1stvamp/hippybot/network

Signed-off-by: Claudio Mignanti <c.mignanti@gmail.com>